02-twitter | refactoring, AuthForm.js 생성 #67
