+++ 02.twitter-clone/nwitter/src/components/AuthForm.js
@@ -1,0 +1,54 @@
+import React, { useState } from 'react';
+import { authService } from 'fbase';
+import { getAuth, createUserWithEmailAndPassword, signInWithEmailAndPassword } from 'firebase/auth';
+
+const AuthForm = () => {
+  const [inputs, setInputs] = useState({
+    email: '',
+    password: '',
+  });
+
+  const { email, password } = inputs;
+  const [newAccount, setNewAccount] = useState(true);
+  const [error, setError] = useState('');
+
+  const onChange = (e) => {
+    const { name, value } = e.target;
+    setInputs({
+      ...inputs,
+      [name]: value,
+    });
+  };
+
+  const onSubmit = async (e) => {
+    e.preventDefault();
+    try {
+      const auth = getAuth();
+      let data;
+      if (newAccount) {
+        data = await createUserWithEmailAndPassword(auth, email, password);
+      } else {
+        data = await signInWithEmailAndPassword(auth, email, password);
+      }
+      console.log(data);
+    } catch (error) {
+      setError(error.message);
+    }
+  };
+
+  const toggleAccount = () => setNewAccount((prev) => !prev);
+
+  return (
+    <>
+      <form onSubmit={onSubmit}>
+        <input name='email' type='text' placeholder='Email' required onChange={onChange} />
+        <input name='password' type='password' placeholder='Password' required onChange={onChange} />
+        <input type='submit' value={newAccount ? 'Create Account' : 'Log In'} />
+        {error}
+      </form>
+      <span onClick={toggleAccount}>{newAccount ? 'Sign In' : 'Create Account'}</span>
+    </>
+  );
+};
+
+export default AuthForm;
